ai-chatbot/pyproject.toml
[project]
name = "ai-chatbot"
version = "0.1.0"
description = "AI Agent - Django ORM Query Generator"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "tortoise-orm[asyncpg]>=0.20.0",
    "aerich>=0.8.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.0",
    "langchain[openai]>=1.0.8",
    "pymupdf>=1.26.6",
    "tqdm>=4.67.1",
    "sentence-transformers==2.6.1",
    "transformers==4.41.2",
    "chromadb>=1.3.5",
    "python-dotenv>=1.2.1",
    "jq>=1.10.0",
    "pysqlite3-binary>=0.5.4",
]
